Implement a faster rejection sampling method.
This method leverages an observation: if we allow ourselves to either sample from a filtered distribution OR the original distribution, we can reduce the amount of data rejected during rejection sampling.

I implemented this 'conditional filtering' in a sort-of hacky way by internally creating an iterator and sticking it in a tf.cond branch.

I also added unit tests to verify that the new sampler works and is faster than the original, and I added benchmarks.

Tested:
bazel test :resample_test